+++ impl-c/include/proc.h
@@ -1,0 +1,8 @@
+#pragma once
+
+#include "proc/task.h"
+
+/* proc/task.c */
+void test_tasks();
+
+void scheduler_init();
